JourneyDashboardView receives wide ApplicationButton as ribbon (#4852)
